requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
argon2==0.1.10
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astroid==3.2.3
astropy==6.1.2
astropy-iers-data==0.2024.7.29.0.32.7
autopep8==2.3.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dill==0.3.8
dnspython==2.6.1
docstring_parser==0.16
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
numpy==2.0.1
orjson==3.10.6
packaging==24.1
parso==0.8.4
passlib==1.7.4
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pyerfa==2.0.1.4
Pygments==2.18.0
pylint==3.2.5
pytest==8.3.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pyuvdata==3.0.0
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
scipy==1.14.0
setuptools==72.1.0
setuptools-scm==8.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
style==1.1.0
tomlkit==0.13.0
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
update==0.0.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
